Move compressor impls to separate packages
